Added temperature polling to the M1. Configure the sensor as described in config_dave.xml. You will need a parameters section after the connection specification in the M1_DEFS to tell the polling thread how often you will want to poll the named sensors.
<PARAMETERS>
	<ITEM NAME="POLL_SENSOR_INTERVAL" VALUE="10"/>
</PARAMETERS>

Each sensor will need to be specified as:
<SENSOR ACTIVE="Y" KEY="01" GROUP="0" NAME="Probe 1" DISPLAY_NAME="PROBE_1" UNITS="C"/>

Where the group can be either 0, 1 or 2. 0 being a probe, 1 being a keypad and 2 being a thermostat.

The flash client will receive a command with the key being the display name and the adjusted temperature in 
Fahrenheit in the extra field.
Group 0 - Subtract 60 from the Temperature Probe data connected to zone temperature data for near minimum -60 
 degrees F. 
Group 1 - Subtract 40 from the Keypad Temperature data. This allows for near minimum -40 degrees F.  
Group 2 - Thermostats send unadjusted temperature.

When the M1 code now starts up it will request the control output status of every connected device. For every 
connected device it will send back to the flash either "on" or "off". The key will be the display name and the
command code will be either "on" or "off".