Have added code to cater for the arming of the M1 from a flash client. For obvious reasons I have not implemented a Disarm function.
In the flash client it will need to send a command like the following:
<CONTROL KEY="ARM" COMMAND="ARM_TO_AWAY" EXTRA="1" EXTRA2="" EXTRA3="" EXTRA4="" EXTRA5=""/>

The key is the ARM and the command can be either:
ARM_TO_AWAY
ARM_TO_STAY_HOME
ARM_TO_STAY_INSTANT
ARM_TO_NIGHT
ARM_TO_NIGHT_INSTANT
ARM_TO_VACATION
ARM_STEP_TO_NEXT_AWAY_MODE
ARM_STEP_TO_NEXT_STAY_MODE

This string must be exact otherwise the command will not be fired.

In the extra field you put the area/partition code (1-8). It will use the user code
(6 digits in length) that is specified in the user config in the command.

In the config file to define a user code you will need the following:
<DEVICE ACTIVE="Y" NAME="M1" DISPLAY_NAME="M1" PASSWORD="123456">

The password would be a user code that could be used only for arming commands and/or other commands
that you would want a flash client to perform.

The Model has been modified so that on startup it creates a Virtual Output device for the ARM command
and adds it to the controlledItem list for the device.