Added the PLC Change Update Command.
The command will come through in the form:
<CONTROL KEY="PLC Change Update" COMMAND="PLC CHANGE UPDATE" EXTRA="<house code>" EXTRA2="<unit code>" EXTRA3="<level>" EXTRA4="" EXTRA5="" />

Where
house code = 'A' to 'P'
unit code = '01' to '16'. If unit code is '00' then it is an 'All Command'
level = 'ON', 'OFF' or the level in % - '02' to '99' unless unit code is an 'All Command' then it can be either 'X10_ALL_UNITS_OFF', 'X10_ALL_LIGHTS_ON' or 'X10_ALL_LIGHTS_OFF'.