Added the following commands to the AutonomicHome - ReportState - StartMCE - SetInstance - StateType - GetStatus
GetStatus will return a series of ReportStates (types defined in the StateType). These will send 
control messages to the client to describe what the extender/MCE box is currently doing.

ControlledHelper is now parsing the ReportState message correctly and updating the in memory 
device for the Windows Media Extender.

Config is parsing the device definition for the WindowsMediaExtender correctly and it has been
added to the DeviceType class.