Added commands to set the current MCE instance. This will set the current instance in the Model.
To set the instance send the following command:
<CONTROL KEY="<key for extender>" COMMAND="SET_INSTANCE" EXTRA="" EXTRA2="" EXTRA3="" EXTRA4="" EXTRA5="" />

As long as the extender is set up in the configuration it will set the current instance on Autonomic
Home to the one selected.

Changed the WindowsMediaExtenderFactory so that it will add each extender as a bi-directional device
(with a message direction of from flash and from the hardware) so that it can accept commands from 
the client. I assume that the key for each extender is unique.