Added a command to Browse for Albums.
To trigger this the flash client sends the following command:
<CONTROL KEY="MCE" COMMAND="BROWSE" EXTRA="ALBUMS" EXTRA2="" EXTRA3="" EXTRA4="" EXTRA5="" />

The Autonomic Home Server will reply with a series of albums (guids and names):
<CONTROL KEY="MCE" COMMAND="ALBUM" EXTRA="{08115859-e625-4bcd-83a8-57e01873b42f}" EXTRA2="Beethoven's Symphony No. 9 (Scherzo) (Seattle Symphony Orchestra)" EXTRA3="" EXTRA4="" EXTRA5="" />

Upon receiving the browse command, BI will attempt to set the instance to the seleced instance (the
key).

TODO: Upon receiving an error inform the client there is an issue.