Added SlimServer as a model. Can browse for albums using the following command <CONTROL KEY="<extender>" COMMAND="BROWSE" EXTRA="<type>" EXTRA2="<start>" EXTRA3="<searchString>" EXTRA4="<artistId>" EXTRA5="<genre>" />
Where type currently only allows ALBUM.
The slim server configuration holds the default number of records to return.
The extender is the as defined in the slim server configuration.
Start specifies at what point we are to start returning information from ie. setting this to 5 will start the result set from the 5th element in the set of results.

artist id and genre are the identifiers of the artists and genres in the slim server. A separate command would be required to return these.

In return, eLife will send back a number of albums as:
<CONTROL KEY="LAPTOP" COMMAND="ALBUM" EXTRA="<album id>" EXTRA2="<album name>" EXTRA3="" EXTRA4="" EXTRA5="" />