The power commands will now return a command with ON or OFF for the particular squeeze device.
PlayerStatus will now return a command with ON or OFF for a particular squeeze device if it is on or off.

Added a new filter to gettracks called currentplaylist that will request PlayerStatus with a given number
of tracks. This will return a certain number of tracks in the current playlist for that player.

The tracks element is now external to the playerstatus similar to the gettracks call.